requirements.txt
async-generator==1.10
attrs==21.4.0
black==22.3.0
blis==0.7.7
catalogue==2.0.7
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
cryptography==36.0.2
cymem==2.0.6
dominate==2.6.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.2.0/en_core_web_sm-3.2.0-py3-none-any.whl
Flask==2.1.1
Flask-Bootstrap4==4.0.2
Flask-WTF==1.0.1
h11==0.13.0
idna==3.3
importlib-metadata==4.11.3
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.1
langcodes==3.3.0
MarkupSafe==2.1.1
murmurhash==1.0.6
mypy-extensions==0.4.3
numpy==1.22.3
outcome==1.1.0
packaging==21.3
pathspec==0.9.0
pathy==0.6.1
pkg_resources==0.0.0
platformdirs==2.5.1
pluggy==1.0.0
preshed==3.0.6
py==1.11.0
pycparser==2.21
pydantic==1.8.2
pyOpenSSL==22.0.0
pyparsing==3.0.7
PySocks==1.7.1
pytest==7.1.1
requests==2.27.1
selenium==4.1.3
smart-open==5.2.1
sniffio==1.2.0
sortedcontainers==2.4.0
spacy==3.2.4
spacy-legacy==3.0.9
spacy-loggers==1.0.2
srsly==2.4.2
# Editable install with no version control (tdd==1.0)
-e /home/fshlinux/hd2/data/programming/python/tdd
thinc==8.0.15
tomli==2.0.1
tqdm==4.63.1
trio==0.20.0
trio-websocket==0.9.2
typer==0.4.1
typing_extensions==4.1.1
urllib3==1.26.9
visitor==0.1.3
wasabi==0.9.1
Werkzeug==2.1.1
wsproto==1.1.0
WTForms==3.0.1
zipp==3.7.0
